requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow==2.14.0
tensorflowjs==4.10.0
tensorflow-hub==0.14.0
h5py==3.8.0
boto3==1.29.3
joblib==1.3.2
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0